requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Gemini API
google-genai>=0.5.0

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0

# Text Processing
markdown>=3.5.0
python-frontmatter>=1.1.0

# SEO & Analytics
textstat>=0.7.0

# Image Processing
Pillow>=10.0.0

# CMS Integration
python-wordpress-xmlrpc>=2.3
notion-client>=2.2.0

# Web Framework (Dashboard)
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-slugify>=8.0.0
tenacity>=8.2.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
